+++ packages/ng/libraries/core/src/lib/event/on-close.subscriber.ts
@@ -1,0 +1,6 @@
+export interface ILuOnCloseSubscriber {
+	onClose(): void;
+}
+export abstract class ALuOnCloseSubscriber implements ILuOnCloseSubscriber {
+	abstract onClose(): void;
+}
