+++ packages/ng/libraries/core/src/lib/event/on-open.subscriber.ts
@@ -1,0 +1,7 @@
+export interface ILuOnOpenSubscriber {
+	onOpen(): void;
+}
+export abstract class ALuOnOpenSubscriber implements ILuOnOpenSubscriber {
+	abstract onOpen(): void;
+}
+
