+++ packages/ng/libraries/core/src/lib/event/on-scroll-bottom.subscriber.ts
@@ -1,0 +1,6 @@
+export interface ILuOnScrollBottomSubscriber {
+	onScrollBottom(): void;
+}
+export abstract class ALuOnScrollBottomSubscriber implements ILuOnScrollBottomSubscriber {
+	abstract onScrollBottom(): void;
+}
