+++ packages/ng/libraries/core/src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './entity/index';
+export * from './event/index';
 export * from './translate/index';
 export * from './tree/index';
